CLN: Clean up core. Remove and rename host.
