+++ src/engramic/application/codify/codify_service.py
@@ -10,13 +10,13 @@ from typing import TYPE_CHECKING, Any
 import tomli
 
 from engramic.core import Engram, Meta, Prompt, PromptAnalysis
+from engramic.core.host import Host
 from engramic.core.metrics_tracker import MetricPacket, MetricsTracker
 from engramic.core.response import Response
 from engramic.core.retrieve_result import RetrieveResult
 from engramic.infrastructure.repository.engram_repository import EngramRepository
 from engramic.infrastructure.repository.meta_repository import MetaRepository
 from engramic.infrastructure.repository.observation_repository import ObservationRepository
-from engramic.infrastructure.system.host_system import HostSystem
 from engramic.infrastructure.system.plugin_manager import PluginManager
 from engramic.infrastructure.system.service import Service
 
@@ -34,7 +34,7 @@ class CodifyService(Service):
     ACCURACY_CONSTANT = 3
     RELEVANCY_CONSTANT = 3
 
-    def __init__(self, host: HostSystem) -> None:
+    def __init__(self, host: Host) -> None:
         super().__init__(host)
         self.plugin_manager: PluginManager = host.plugin_manager
         self.engram_repository: EngramRepository = EngramRepository(self.plugin_manager)
