+++ src/engramic/infrastructure/system/__init__.py
@@ -1,2 +1,1 @@
-from engramic.infrastructure.system.host_system import HostSystem as HostSystem
 from engramic.infrastructure.system.service import Service as Service
